Привела проект к правильной структуре
