--- address-book-tests/src/test/java/ru/rt/client/tests/ContactTest.java
@@ -1,17 +1,0 @@
-package ru.rt.client.tests;
-
-import org.testng.annotations.Test;
-import ru.rt.client.model.CompanyDate;
-import ru.rt.client.model.Name;
-import ru.rt.client.model.PhoneNumber;
-
-public class ContactTest extends TestBase {
-
-
-    @Test
-    public void testContact() throws Exception {
-        app.nevContact();
-        app.fillContact(new Name("Ivan", "Ivanovich", "Ivanov", "Vano"), new CompanyDate("Ivanov-and-K", "Moscow"), new PhoneNumber("8(861)334-33-33", "8(927)333-33-33", "8(861)444-44-44", "8(861)444-44-45"), "vano@bk.ru");
-        app.submitContact();
-    }
-}
