+++ src/test/java/ru/rt/client/tests/GroupDeletionTest.java
@@ -1,0 +1,13 @@
+package ru.rt.client.tests;
+
+import org.testng.annotations.Test;
+
+public class GroupDeletionTest extends TestBase {
+
+    @Test
+    public void testGroupDeletion() throws Exception {
+        app.navToGroupsPage();
+        app.selectGroup();
+        app.clickDeleteBtn();
+    }
+}
